+++ crates/router/src/events.rs
@@ -9,6 +9,7 @@ pub mod api_logs;
 pub mod connector_api_logs;
 pub mod event_logger;
 pub mod kafka_handler;
+pub mod outgoing_webhook_logs;
 
 pub(super) trait EventHandler: Sync + Send + dyn_clone::DynClone {
     fn log_event(&self, event: RawEvent);
